% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camateurclass21-package.R
\docType{package}
\name{camateurclass21-package}
\alias{camateurclass21}
\alias{camateurclass21-package}
\title{camateurclass21: Title}
\description{
..
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com}

}
\keyword{internal}
